Refactor firewall check.
