+++ agent/security_helper.py
@@ -1,4 +1,3 @@
-from unittest import mock
 import iptc
 from xml.etree import ElementTree as ET
 from sh import nmap
@@ -27,24 +26,6 @@ def nmap_scan(target):
                 'state': state
             })
     return result
-
-
-def test_firewall_enabled_pos():
-    with mock.patch('iptc.Table') as ipt:
-        chain0 = mock.Mock()
-        chain0.name = 'INPUT'
-        chain0.rules = [object(), object()]
-        ipt.return_value = [chain0]
-        assert is_firewall_enabled() is True
-
-
-def test_firewall_enabled_neg():
-    with mock.patch('iptc.Table') as ipt:
-        chain0 = mock.Mock()
-        chain0.name = 'INPUT'
-        chain0.rules = []
-        ipt.return_value = [chain0]
-        assert is_firewall_enabled() is False
 
 
 def is_firewall_enabled():
